# refactor: convert IstioOperator from Unstructured more efficiently