--- a/istioctl/pkg/install/verify_test.go
+++ b/istioctl/pkg/install/verify_test.go
@@ -17,15 +17,11 @@ package install
 import (
 	"fmt"
 	"testing"
-	"time"
 
 	appsv1 "k8s.io/api/apps/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 
 	"istio.io/api/operator/v1alpha1"
-
-	operatorv1alpha1 "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 )
 
 var (
@@ -102,37 +98,38 @@ var (
 		},
 	}
 
-	sampleIOP = &operatorv1alpha1.IstioOperator{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "IstioOperator",
-			APIVersion: "install.istio.io/v1alpha1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "installed-state",
-			Namespace: "istio-system",
-			ManagedFields: []metav1.ManagedFieldsEntry{
-				{
-					FieldsType: "FieldsV1",
-					FieldsV1: &metav1.FieldsV1{
-						Raw: []byte(`{"f:metadata": {}}`),
+	sampleIOP = unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"kind":       "IstioOperator",
+			"apiVersion": "install.istio.io/v1alpha1",
+			"metadata": map[string]interface{}{
+				"name":              "installed-state",
+				"namespace":         "istio-system",
+				"creationTimestamp": "2020-07-29T07:08:55Z",
+				"managedFields": []map[string]interface{}{
+					{
+						"fieldsType": "FieldsV1",
+						"fieldsV1": map[string]interface{}{
+							"f:metadata": map[string]interface{}{},
+						},
+						"manager":   "istioctl",
+						"operation": "Update",
+						"time":      "2020-07-29T07:09:05Z",
 					},
-					Manager:   "istioctl",
-					Operation: metav1.ManagedFieldsOperationUpdate,
-					Time:      &metav1.Time{Time: time.Now()},
 				},
 			},
-			CreationTimestamp: metav1.Time{Time: time.Now()},
-		},
-		Spec: &v1alpha1.IstioOperatorSpec{Profile: "default"},
-		Status: &v1alpha1.InstallStatus{
-			Status: v1alpha1.InstallStatus_HEALTHY,
-			ComponentStatus: map[string]*v1alpha1.InstallStatus_VersionStatus{
-				"Base": {
-					Status: v1alpha1.InstallStatus_HEALTHY,
+			"spec": map[string]interface{}{
+				"profile": "default",
+			},
+			"status": map[string]interface{}{
+				"status": "HEALTHY",
+				"componentStatus": map[string]interface{}{
+					"Base": map[string]interface{}{
+						"status": "HEALTHY",
+					},
 				},
 			},
-		},
-	}
+		}}
 )
 
 func TestGetDeploymentStatus(t *testing.T) {
@@ -238,12 +235,7 @@ func TestFindResourceInSpec(t *testing.T) {
 }
 
 func TestIstioOperatorConversion(t *testing.T) {
-	unObj, err := runtime.DefaultUnstructuredConverter.ToUnstructured(sampleIOP)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	out, err := convertToIstioOperator(unObj)
+	out, err := convertToIstioOperator(sampleIOP.Object)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -254,14 +246,8 @@ func TestIstioOperatorConversion(t *testing.T) {
 }
 
 func BenchmarkConvertIstioOperator(b *testing.B) {
-	unObj, err := runtime.DefaultUnstructuredConverter.ToUnstructured(sampleIOP)
-	if err != nil {
-		b.Fatal(err)
-	}
-
-	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		_, err := convertToIstioOperator(unObj)
+		_, err := convertToIstioOperator(sampleIOP.Object)
 		if err != nil {
 			b.Fatal(err)
 		}