Signed-off-by: knight42 <anonymousknight96@gmail.com>

Currently we convert `IstioOperator` from `Unstructured` by marshaling it to YAML first(`util.ToYAML`) then converting it to JSON(`yaml.YAMLToJSON` in `util.UnmarshalWithJSONPB`) and finally unmarshaling from JSON(`jsonpb.Unmarshaler`). Given that we already know what its type is, we could leverage `runtime.DefaultUnstructuredConverter.FromUnstructured` to convert `Unstructured` to `IstioOperator` directly.

In addition, it also brings us other benefits:

1. eliminates subtle bugs that deserialization of `IstioOperator` chokes on `metav1.Time`. It is worth noting that `metadata.managedFields` is introduced in k8s 1.18, which includes `metav1.Time` in [ManagedFieldsEntry](https://pkg.go.dev/k8s.io/apimachinery/pkg/apis/meta/v1?tab=doc#ManagedFieldsEntry), and it will make `istioctl verify-install` complain "Error: could not load IstioOperator from cluster: could not unmarshal: json: cannot unmarshal string into Go value of type map[string]json.RawMessage". #25688 

2. impressive performance improvement.

Before:
```
2020-07-28T12:45:53.414688Z	info	proto: tag has too few fields: "-"
goos: darwin
goarch: amd64
pkg: istio.io/istio/istioctl/pkg/install
BenchmarkConvertIstioOperator-16    	    4808	    215618 ns/op	  126765 B/op	    1160 allocs/op
PASS
ok  	istio.io/istio/istioctl/pkg/install	2.189s
```

After:
```
goos: darwin
goarch: amd64
pkg: istio.io/istio/istioctl/pkg/install
BenchmarkConvertIstioOperator-16    	   31864	     37360 ns/op	   13651 B/op	     236 allocs/op
PASS
ok  	istio.io/istio/istioctl/pkg/install	2.362s
```

benchcmp
```
benchmark                            old ns/op     new ns/op     delta
BenchmarkConvertIstioOperator-16     215618        37360         -82.67%

benchmark                            old allocs     new allocs     delta
BenchmarkConvertIstioOperator-16     1160           236            -79.66%

benchmark                            old bytes     new bytes     delta
BenchmarkConvertIstioOperator-16     126765        13651         -89.23%
```

And to help us figure out who should review this PR, please 
put an X in all the areas that this PR affects.

[ ] Configuration Infrastructure
[ ] Docs
[x] Installation
[ ] Networking
[x] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[x] User Experience
[ ] Developer Infrastructure