@istio-testing 2020-07-28
@knight42: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
release-notes_istio | fbf3c9d00a37f4b8148ef85b23273eec2407dcb8 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/25913/release-notes_istio/764) | `/test release-notes_istio`
lint_istio | fbf3c9d00a37f4b8148ef85b23273eec2407dcb8 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/25913/lint_istio/17004) | `/test lint_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->