@istio-testing 2020-07-28
@knight42: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
gencheck_istio | 59d332ff1b253dafc9b4d63129bb644a595f6344 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/25913/gencheck_istio/15590) | `/test gencheck_istio`
integ-ipv6-k8s-tests_istio | 59d332ff1b253dafc9b4d63129bb644a595f6344 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/25913/integ-ipv6-k8s-tests_istio/3274) | `/test integ-ipv6-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->