@sdake 2020-07-28
@knight42 in theory - you got  rid of the warning `2020-07-28T12:45:53.414688Z	info	proto: tag has too few fields: "-"
`. 

@morvencao would you mind taking a look at this PR and provide your thoughts?

Cheers,
-steve