@knight42 2020-07-28
Furthermore, as for long term I believe we need a typed client to manage `IstioOperator` so that we don't need such conversion in each place we need a `IstioOperator`.