@knight42 2020-07-29
@morvencao I think it makes sense, but we actually get the unstructured object from dynamic client rather than yaml/json, how about I create the unstructured object literal directly?