@tanjunchen 2020-07-29
> Furthermore, as for long term I believe we need a typed client to manage `IstioOperator` so that we don't need such manual conversion in each place we need a `IstioOperator`.

+1  client  for IstioOperator