@morvencao 2020-07-29
@knight42 
> how about I create the unstructured object literal directly?
OK to me